Add QueuedStatementResource timeout for query submission (mem leak fix)
QueuedStatementResource attempts to maintain an in-memory query cache by
only expiring queries that are no longer tracked by DispatchManager. It
is possible for a query to be created in QueuedStatementResource, but
never submitted to DispatchManager if the client never comes back and
calls getStatus() (which is the trigger for the query to be submitted to
DispatchManager).

To handle this, we add a timeout to purge queries from this cache if it
has not been submitted by the query abandoned client timeout threshold.

# Conflicts:
#	presto-main/src/main/java/io/prestosql/dispatcher/QueuedStatementResource.java